Disable SSD journals if only one disk type available (bnc#916440, bnc#918831)
If a node has only SSDs, or only HDDs, attempting automatic journal
allocation to SSDs makes no sense.  Also, in the case of a system with
only SSDs, the previous implementation would allocate them all as
journals with no actual OSDs, which is obviously broken.

This also fixes the edge case where if a node only has one disk
available it was being allocated as a journal with no OSD.

Signed-off-by: Tim Serong <tserong@suse.com>