Only use disks bigger than 20GB (bnc#886795)
Disks 10GB or less result in weighting errors.  5GB is used for the
journal by default, so 20GB seems the safest minium to run with.  Note
that real-world deployments will be unlikely to attempt to use disks
this small for OSDs, but it's not uncommon to run into problems with
small disks when testing toy clusters on virtual machines, which is why
we're now enforcing a minimum size.

Signed-off-by: Tim Serong <tserong@suse.com>