nova: Correctly handle changing libvirt secrets (bsc#923224)
It might happen that a secret defined in libvirt blocks us from setting
the secret we need (because the usage name is already taken, or because
the usage name is different for the same UUID). If that's the case, then
we undefine that conflicting secret first.

Otherwise, we'd get an error defining our secret.

https://bugzilla.suse.com/show_bug.cgi?id=923224