Add support for deploying Calamari with HTTPS (bsc#915315)
Note that this change causes the barclamp to replace
/etc/apache2/conf.d/calamari.conf as shipped in the calamari-server
pacakge.  This should be fine as that file is %config(noreplace) in the
calamari-server specfile, so won't be clobbered if calamari-server is
upgraded later.  The user is expected to generate suitable certificates,
or rely on default certificates to be generated by the calamari-server
package in %post.

Signed-off-by: Tim Serong <tserong@suse.com>