radosgw: Create limited permission ceph-radosgw.conf (bsc#958470)
/etc/ceph/ceph.conf needs to be chmod 644, because various services need
to be able to read the file.  When deploying radosgw with keystone, the
keystone admin token ordinarily ends up in ceph.conf, and being
world-readable that's a potential security risk.

This change adds /etc/ceph/ceph-radosgw.conf, which is almost identical
to the regular ceph.conf, but chmod 640, and including the keystone
settings.  It also generates /etc/systemd/system/ceph-radosgw@.service
so that radosgw is started using the new config file.

Signed-off-by: Tim Serong <tserong@suse.com>