Handle down nodes when searching for MONs and OSDs (bsc#972001)
Searching for OSDs using the query "run_list:role\[ceph-osd\]" works
fine if all the nodes are up, but if any OSD nodes are down, they're not
returned by this search (somehow the ceph-osd role disappears from the
role run list when the node goes down, and reappears when it comes back
up).  This can lead to a case where a node which is a ceph client hits a
ZeroDivisionError applying the ceph recipe if it comes up before any
OSDs are available (OSD count is zero in this case, at least
transiently).

This commit changes the search to query "run_list_map:ceph_osd", which
seems to work fine regardless of whether nodes are up or down.

Signed-off-by: Tim Serong <tserong@suse.com>