Add support for deploying Ceph MDS (fate#321033)
This allows deploying up to three Ceph MDS servers.  MDSes should
optimally be run on their own hosts, but can be colocated with MONs
on smaller installations.  They cannot be colocated with OSDs.

The filesystem is named cephfs, and uses the cephfs_metadata and
cephfs_data pools, which will be created automatically.  These
names are defined in attributes so could be overriden if necessary,
but are not exposed in the UI.

If more than one MDS is deployed, by default they will *not* run
multi-active (that would require enabling some extra options).

Signed-off-by: Tim Serong <tserong@suse.com>