Put ceph client traffic on public network (fate#319122)
For new deploys, this defaults ceph client traffic to the public
network, which is necessary for using Manila with CephFS.  If you
don't want the ceph client traffic to be on the public network,
edit the Ceph proposal in raw mode and set "client_network" to
some other network.

When the migration runs to add the client_network attribute, it
will default this value to "admin", so that existing deployments
keep their client traffic on the admin network (if this changed
automatically duing an upgrade it would cause all sorts of
trouble -- it'd mean needing to change the MON IP addresses,
which you're really, really not meant to do).

The one caveat is that if you have an existing deployment, then
upgrade to this version and deploy an MDS, the client traffic
is still on the admin network, which AIUI means you won't be
able to use Manila with CephFS on an upgraded deployment, because
the client traffic is on a network inaccessible to cloud instances.

Signed-off-by: Tim Serong <tserong@suse.com>