get_osd_nodes(): just return hostnames, not net config (bsc#1015069)
AFAIK nothing is actually using the cluster and public network info
returned by this function, and trying to do so causes problems: AFAICT
the attribute node["ceph"]["client_network"] is only set after the ceph
roles are applied, so if you have an OSD which isn't also a MON (normal
in anything other than a toy test cluster), while the MONs are being
configured, get_osd_nodes() is called, and because the OSD nodes don't
have that attribute set yet, this will fail.  So, let's just delete
this piece.

Signed-off-by: Tim Serong <tserong@suse.com>