ceph: Fix integration with external ceph clusters (bsc#1025674)
With an external ceph cluster, a chef search for mons will result in no
result, and that means we're not able to get the network that is used
for ceph. And we're crashing.

But with external ceph clusters, this information is only used to decide
which hostname to use (since we pick a hostname based on the network),
and that's actually irrelevant: the value for the hostname actually only
really matters for nodes that are part of the ceph clusters, not for
client nodes -- and the code is only used for client nodes when we're
using an external ceph cluster.

https://bugzilla.suse.com/show_bug.cgi?id=1025674