Update RGW HA deployment to use systemd services (bsc#956582)
RGW HA deployments are presently broken; it tries to make a
clone of a ceph-radosgw service, but this doesn't work, because
we've long since been using parameterized systemd services in
the form ceph-radosgw@rgw.$(hostname).  This commit removes
the clone and replaces it with individual primitives for each
host with appropriate location constraints.

Signed-off-by: Tim Serong <tserong@suse.com>