Better selection for a nodes with ceph-mds and ceph-calamari roles.
We cannot use standard select_nodes_for_role method for ceph-mds role
because it limits the search for available nodes to maximum of 'count',
which is a number that is defined in role_constraints for each role.
We also don't want to remove that count for ceph-mds, because is
also used to limit user when manually assigning role.

Also, make sure ceph-calamari is not deployed to cluster node.