radosgw: Generate slightly saner root RSA key
The default root key of 4096 RSA bits is not very performant,
apparently openssl isn't able to optimize that. Since all the
rest of the self generated certs use 2048, and that should be
secure for a while (especially for a service that does not setup
a valid cert and that isn't being checked by the client) that should
be plenty enough.

Improves connection performance by factor (!) 2.5.